requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
SQLAlchemy==2.0.23
WTForms==3.1.1
PyJWT==2.8.0
bcrypt==4.1.2
gunicorn==21.2.0 